Restructure the EKS deployment guide
There are a few back and forths in the EKS deployment guide, which can
make it harder to follow. This will move the hostname and TLS configuration
before the gitlab application creation, so there's no back and forth between
these sections, making it easier to follow.

Splits the TLS configuration section in 2. Moves the "own certificates" section
in the second part.

Move the "/etc/hosts" and Amazon Load Balancer connection methods to the Appendix.