actions: Adds action which checks the jobs status on related projects
The Legend Juju Bundle has quite a few components spread across several
repositories, which can make it a bit harder to track and ensure that everything
is healthy.

This adds an action which will check the status of all the jobs of all the related
projects, making it easier to maintain.