Adds EKS deployment Legend Prod and Staging switch Github Action
We currently have 2 Juju FINOS Legend models in the EKS deployment,
finos-legend and finos-legend-twin, one of which will be considered
the Production environment and the other one as the Staging environment.
The plan is to keep the Staging environment up-to-date with the latest
Legend releases, and switch the environments monthly.

Adds a dispatchable GitHub action that will switch these 2 environments.