actions: Update the actions for the staging model
The TLS certificate needs to be refreshed for both finos-legend
and finos-legend-twin models.

For the Juju refresh action, it has been decided to only keep the
staging model up-to-date and switch which of the models will become
the prod model or the staging model. For this, we need to check which
model is the staging one.