actions: Disable Refresh EKS Legend deployment on switch
After switching the Legend environment staging and Prod environments,
disable the Refresh EKS FINOS Legend deployment, so we have the
ability to switch back if something went wrong with the new Prod
environment. Otherwise, the daily action would refresh the staging
environment to the latest version.

The disabled action will have to be enabled manually afterwards.